Moved all code in SearchBar component into HomePage component. Deleted files in SearchBar component. This was done to try and solve the problem from another perspective.
