--- src/components/SearchProductBar/index.styles.ts
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export const Input = styled.input`
-  background: gray;
-  border: solid 2px black;
-`;
