--- src/components/SearchProductBar/index.tsx
@@ -1,18 +1,0 @@
-import * as React from "react";
-import * as Styles from "./index.styles";
-
-const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-  const userSearchInputValue = event.target.value.trim().toLowerCase();
-};
-
-function SearchProductBar() {
-  return (
-    <>
-      <section>
-        <Styles.Input onChange={handleInputChange} />
-        <button>Search</button>
-      </section>
-    </>
-  );
-}
-export default SearchProductBar;
